+++ ha-business/src/main/java/jp/co/ha/business/db/crud/read/AccountRecoveryTokenSearchService.java
@@ -26,6 +26,6 @@ public interface AccountRecoveryTokenSearchService {
      * @return アカウント回復トークン
      */
     Optional<AccountRecoveryTokenData> findBySeqUserIdAndTokenAndValidTokenCreateDate(
-            Integer seqUserId, String token);
+            Long seqUserId, String token);
 
 }
